test: Fix sqs docker container test
