+++ springwolf-examples/springwolf-sqs-example/docker-compose.yml
@@ -4,6 +4,8 @@ services:
     image: stavshamir/springwolf-sqs-example:${SPRINGWOLF_VERSION}
     links:
       - localstack
+    environment:
+      - SPRING_CLOUD_AWS_ENDPOINT=http://localstack:4566
     ports:
       - "8080:8080"
     depends_on:
@@ -15,9 +17,9 @@ services:
       - DEBUG=${DEBUG-}
       - AWS_REGION=eu-central-1
       - SERVICES=sqs,sns
-    ports:
-      - "4566:4566"            # LocalStack Gateway
-      - "4510-4559:4510-4559"  # external services port range
+#    ports:
+#      - "4566:4566"            # LocalStack Gateway
+#      - "4510-4559:4510-4559"  # external services port range
     volumes:
       - "${TMPDIR:-/tmp/localstack}:/tmp/localstack"
       - "/var/run/docker.sock:/var/run/docker.sock"
@@ -30,5 +32,5 @@ services:
     restart: "no"
     entrypoint: [ "bash", "-c", "
       awslocal --endpoint-url=http://localstack:4566 sqs create-queue --queue-name another-queue --region eu-central-1;
-      awslocal --endpoint-url=http://localstack:4566 sqs create-queue --queue-name example-queue --region eu-central-1
+      awslocal --endpoint-url=http://localstack:4566 sqs create-queue --queue-name example-queue --region eu-central-1;
     " ]
